flaskapi/requirements.txt
flask
pydash
python-dotenv
openai
aiohttp
requests
pymongo
openai-whisper